+++ eskema/ddlgen/monkey.py
@@ -3,7 +3,7 @@ import ddlgenerator.reshape
 import ddlgenerator.typehelpers
 
 from .ddlgenerator import AnyDialect, TablePlus
-from .reshape import clean_key_name
+from .reshape import clean_key_name, use_this_pk
 from .typehelpers import best_coercable, coerce_to_specific
 
 
@@ -12,5 +12,6 @@ def activate():
     ddlgenerator.ddlgenerator.mock_engines = AnyDialect()
     ddlgenerator.ddlgenerator.Table = TablePlus
     ddlgenerator.reshape.clean_key_name = clean_key_name
+    ddlgenerator.reshape.ParentTable.use_this_pk = use_this_pk
     ddlgenerator.typehelpers.best_coercable = best_coercable
     ddlgenerator.typehelpers.coerce_to_specific = coerce_to_specific