+++ eskema/ddlgen/reshape.py
@@ -1,5 +1,5 @@
 import sqlalchemy as sa
-from ddlgenerator.reshape import _illegal_in_column_name
+from ddlgenerator.reshape import UniqueKey, _illegal_in_column_name, all_values_for
 
 
 def clean_key_name(key: str) -> str:
@@ -34,3 +34,19 @@ def clean_key_name(key: str) -> str:
     # Patch: Just to make sure?
     result = sa.sql.quoted_name(result, quote='"')
     return result
+
+
+def use_this_pk(self, pk_name, key_type):
+    """
+    More graceful version which will ignore errors on `int` type PKs,
+    and fall back to the effective type instead.
+    """
+    if key_type == int:
+        try:
+            all_max = max([0] + all_values_for(self, pk_name))
+            self.pk = UniqueKey(pk_name, key_type, max=all_max)
+            return
+        except TypeError:
+            pass
+
+    self.pk = UniqueKey(pk_name, key_type)